refactor: package 네이밍 및 구조 변경
