+++ src/main/kotlin/at/voyagers/common/order/application/OrderService.kt
@@ -1,5 +1,6 @@
 package at.voyagers.common.order.application
 
+import at.voyagers.common.order.application.component.OrderManager
 import at.voyagers.common.order.domain.Order
 import at.voyagers.common.order.domain.OrderCreateSpec
 import org.springframework.stereotype.Service
