--- src/main/kotlin/at/voyagers/common/order/domain/OrderRepository.kt
@@ -1,7 +1,0 @@
-package at.voyagers.common.order.domain
-
-interface OrderRepository {
-
-    fun save(order: Order): Order
-
-}
