--- src/main/kotlin/at/voyagers/common/order/domain/OrderVariantRepository.kt
@@ -1,4 +1,0 @@
-package at.voyagers.common.order.domain
-
-interface OrderVariantRepository {
-}
