+++ src/main/kotlin/at/voyagers/common/order/domain/repository/OrderRepository.kt
@@ -1,0 +1,9 @@
+package at.voyagers.common.order.domain.repository
+
+import at.voyagers.common.order.domain.Order
+
+interface OrderRepository {
+
+    fun save(order: Order): Order
+
+}
