+++ src/main/kotlin/at/voyagers/common/order/domain/repository/OrderVariantRepository.kt
@@ -1,0 +1,4 @@
+package at.voyagers.common.order.domain.repository
+
+interface OrderVariantRepository {
+}
